added the class notes
I took these class notes .. and this first class, before deciding to use
github to document my work.  That was my mistake.

Each class that offers no certification in the future should have a
project on github with notes, screenshots of the syllablus in the wiki,
and a write of the what the course covered.

I'll be adding those sorts of details here in the near future.